requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
pillow>=9.0.0
numpy>=1.21.0
pydantic>=2.0.0
python-multipart>=0.0.6
accelerate>=0.24.0
Pillow==10.1.0
requests==2.31.0 